project_name/requirements.txt
flask-restplus==0.13.0
Flask-SQLAlchemy==2.1
Flask-Cors==3.0.7
Werkzeug==0.16.0
loguru==0.4.1

numpy==1.18.4
Pillow==7.1.2
